+++ extensions/raycast-ollama/src/api/types.ts
@@ -1,8 +1,29 @@
+export interface OllamaApiTagsResponse {
+  models: OllamaApiTagsResponseModel[];
+}
+
+export interface OllamaApiTagsResponseModel {
+  name: string;
+  modified_at: string;
+  size: number;
+  download?: number;
+}
+
+export interface OllamaApiPullResponse {
+  status: string;
+  digest?: string;
+  total?: number;
+  completed?: number;
+  error?: string;
+}
+
 export interface OllamaApiGenerateRequestBody {
   model: string;
   prompt: string;
   system?: string;
   template?: string;
+  context?: number[];
+  options?: OllamaApiGenerateOptionsRequestBody;
 }
 
 export interface OllamaApiGenerateOptionsRequestBody {
@@ -63,12 +84,6 @@ export interface OllamaApiEmbeddingsResponse {
   embedding: number[];
 }
 
-export interface OllamaApiTagsResponseModel {
-  name: string;
-  modified_at: string;
-  size: string;
-}
-
 export interface OllamaApiGenerateResponse {
   metadata: OllamaApiGenerateResponseMetadata;
   answer: string;
@@ -107,8 +122,17 @@ export interface RaycastArgumentsOllamaAsk {
 export interface RaycastArgumentsOllamaAskCustom {
   fallbackText?: string;
   arguments: {
-    prompt: string;
+    model: string;
     query: string;
+  };
+  launchType: string;
+  launchContext?: string;
+}
+
+export interface RaycastArgumentsOllamaChatCustom {
+  fallbackText?: string;
+  arguments: {
+    model: string;
   };
   launchType: string;
   launchContext?: string;
@@ -117,8 +141,7 @@ export interface RaycastArgumentsOllamaAskCustom {
 export interface RaycastArgumentsOllamaCommandCustom {
   fallbackText?: string;
   arguments: {
-    prompt: string;
-    query: string;
+    model: string;
   };
   launchType: string;
   launchContext?: string;