+++ extensions/raycast-ollama/src/ollama-custom-command.tsx
@@ -1,13 +1,9 @@
 import { OllamaApiGenerateRequestBody, RaycastArgumentsOllamaCommandCustom } from "./api/types";
-import ResultView from "./api/main";
-import { getPreferenceValues } from "@raycast/api";
-
-const preferences = getPreferenceValues();
+import { ResultView } from "./api/main";
 
 export default function Command(props: RaycastArgumentsOllamaCommandCustom): JSX.Element {
   const body = {
-    model: preferences.ollamaAskCustomCommandModel,
-    system: props.arguments.prompt,
+    model: props.arguments.model,
   } as OllamaApiGenerateRequestBody;
 
   return ResultView(body, true);