+++ extensions/raycast-ollama/CHANGELOG.md
@@ -1,5 +1,11 @@
 # raycast-ollama Changelog
 
+## [Improvement] - 2023-08-15
+
+- New command **Chat with Ollama**: an interactive chat (request / response) with possibility to save multiple conversation and resume them.
+- New command **Create Custom Command**: an easy way for create custom commands with quicklinks.
+- New command **Manage Models**: where you can view installed models, delete them and install new ones.
+
 ## [Improvement and BugFix] - 2023-08-12
 
 - No longer needed custom MODFILEs.